requirements.txt
pyyaml
mypy
types-PyYAML
pytest
